Remove generate plots for process-export-files and fix unit test
Remove creating plot directory
